refactor: delete HistoryDisplay component
